CalendarPlanner: initial commit
Roughly put together, the purpose of this project is to provide a means
to play around with Calendar schedules without actually modifying the
real calendar.

This project uses client-side google API calls which allows the user to
explicitly grant authorization and see that their calendar is only being
rendered within their browser. The API scope is read only.

Next steps for this will be to cleanup and productionize this code so
that people (myself included) can use this to rearrange their workday
schedules... in theory!