Add ical client-side upload and parsing
If a user does not have permissions to / does not want to provide access
to a google calendar, this allows them to instead upload (client-side)
an iCal format calendar that will be parsed and loaded into the calendar

What needs to be fixed: recurring meetings are not showing up correctly.
This needs to be fixed.