client.js: parse dtstart for yearly events
yearly events DTSTART format is not parsed properly by the ical library and results
in the start date for the recurring event being the current day.

This commit takes the DTSTART value for the event and parses it into the
RRULE which allows it to recurr on the correct day.

The underlying parsing library could be fixed to address this. Currently
these yearly events have a set timezone which may cause the events to
appear a day early.